Move code around to make common writer
