+++ bio2zarr/writer.py
@@ -1,0 +1,967 @@
+import dataclasses
+import json
+import logging
+import os
+import pathlib
+import shutil
+
+import numcodecs
+import numpy as np
+import zarr
+
+from bio2zarr import constants, core, provenance, schema, zarr_utils
+from bio2zarr.vcf2zarr import icf, vcz
+
+logger = logging.getLogger(__name__)
+
+
+@dataclasses.dataclass
+class LocalisableFieldDescriptor:
+    array_name: str
+    vcf_field: str
+    sanitise: callable
+    convert: callable
+
+
+def compute_lad_field(ad, la):
+    assert ad.shape[0] == la.shape[0]
+    assert la.shape[1] == 2
+    lad = np.full((ad.shape[0], 2), -2, dtype=ad.dtype)
+    homs = np.where((la[:, 0] != -2) & (la[:, 1] == -2))
+    lad[homs, 0] = ad[homs, la[homs, 0]]
+    hets = np.where(la[:, 1] != -2)
+    lad[hets, 0] = ad[hets, la[hets, 0]]
+    lad[hets, 1] = ad[hets, la[hets, 1]]
+    return lad
+
+
+def pl_index(a, b):
+    """
+    Returns the PL index for alleles a and b.
+    """
+    return b * (b + 1) // 2 + a
+
+
+def compute_lpl_field(pl, la):
+    lpl = np.full((pl.shape[0], 3), -2, dtype=pl.dtype)
+
+    homs = np.where((la[:, 0] != -2) & (la[:, 1] == -2))
+    a = la[homs, 0]
+    lpl[homs, 0] = pl[homs, pl_index(a, a)]
+
+    hets = np.where(la[:, 1] != -2)[0]
+    a = la[hets, 0]
+    b = la[hets, 1]
+    lpl[hets, 0] = pl[hets, pl_index(a, a)]
+    lpl[hets, 1] = pl[hets, pl_index(a, b)]
+    lpl[hets, 2] = pl[hets, pl_index(b, b)]
+
+    return lpl
+
+
+localisable_fields = [
+    LocalisableFieldDescriptor(
+        "call_LAD", "FORMAT/AD", icf.sanitise_int_array, compute_lad_field
+    ),
+    LocalisableFieldDescriptor(
+        "call_LPL", "FORMAT/PL", icf.sanitise_int_array, compute_lpl_field
+    ),
+]
+
+
+@dataclasses.dataclass
+class VcfZarrPartition:
+    start: int
+    stop: int
+
+    @staticmethod
+    def generate_partitions(num_records, chunk_size, num_partitions, max_chunks=None):
+        num_chunks = int(np.ceil(num_records / chunk_size))
+        if max_chunks is not None:
+            num_chunks = min(num_chunks, max_chunks)
+        partitions = []
+        splits = np.array_split(np.arange(num_chunks), min(num_partitions, num_chunks))
+        for chunk_slice in splits:
+            start_chunk = int(chunk_slice[0])
+            stop_chunk = int(chunk_slice[-1]) + 1
+            start_index = start_chunk * chunk_size
+            stop_index = min(stop_chunk * chunk_size, num_records)
+            partitions.append(VcfZarrPartition(start_index, stop_index))
+        return partitions
+
+
+VZW_METADATA_FORMAT_VERSION = "0.1"
+
+
+@dataclasses.dataclass
+class VcfZarrWriterMetadata(core.JsonDataclass):
+    format_version: str
+    icf_path: str
+    schema: schema.VcfZarrSchema
+    dimension_separator: str
+    partitions: list
+    provenance: dict
+
+    @staticmethod
+    def fromdict(d):
+        if d["format_version"] != VZW_METADATA_FORMAT_VERSION:
+            raise ValueError(
+                "VcfZarrWriter format version mismatch: "
+                f"{d['format_version']} != {VZW_METADATA_FORMAT_VERSION}"
+            )
+        ret = VcfZarrWriterMetadata(**d)
+        ret.schema = schema.VcfZarrSchema.fromdict(ret.schema)
+        ret.partitions = [VcfZarrPartition(**p) for p in ret.partitions]
+        return ret
+
+
+@dataclasses.dataclass
+class VcfZarrWriteSummary(core.JsonDataclass):
+    num_partitions: int
+    num_samples: int
+    num_variants: int
+    num_chunks: int
+    max_encoding_memory: str
+
+
+class VcfZarrWriter:
+    def __init__(self, path):
+        self.path = pathlib.Path(path)
+        self.wip_path = self.path / "wip"
+        self.arrays_path = self.wip_path / "arrays"
+        self.partitions_path = self.wip_path / "partitions"
+        self.metadata = None
+        self.icf = None
+
+    @property
+    def schema(self):
+        return self.metadata.schema
+
+    @property
+    def num_partitions(self):
+        return len(self.metadata.partitions)
+
+    def has_genotypes(self):
+        for field in self.schema.fields:
+            if field.name == "call_genotype":
+                return True
+        return False
+
+    def has_local_alleles(self):
+        for field in self.schema.fields:
+            if field.name == "call_LA" and field.vcf_field is None:
+                return True
+        return False
+
+    #######################
+    # init
+    #######################
+
+    def init(
+        self,
+        icf,
+        *,
+        target_num_partitions,
+        schema,
+        dimension_separator=None,
+        max_variant_chunks=None,
+    ):
+        self.icf = icf
+        if self.path.exists():
+            raise ValueError("Zarr path already exists")  # NEEDS TEST
+        schema.validate()
+        partitions = VcfZarrPartition.generate_partitions(
+            self.icf.num_records,
+            schema.variants_chunk_size,
+            target_num_partitions,
+            max_chunks=max_variant_chunks,
+        )
+        # Default to using nested directories following the Zarr v3 default.
+        # This seems to require version 2.17+ to work properly
+        dimension_separator = (
+            "/" if dimension_separator is None else dimension_separator
+        )
+        self.metadata = VcfZarrWriterMetadata(
+            format_version=VZW_METADATA_FORMAT_VERSION,
+            icf_path=str(self.icf.path),
+            schema=schema,
+            dimension_separator=dimension_separator,
+            partitions=partitions,
+            # Bare minimum here for provenance - see comments above
+            provenance={"source": f"bio2zarr-{provenance.__version__}"},
+        )
+
+        self.path.mkdir()
+        root = zarr.open(store=self.path, mode="a", **zarr_utils.ZARR_FORMAT_KWARGS)
+        root.attrs.update(
+            {
+                "vcf_zarr_version": "0.2",
+                "vcf_header": self.icf.vcf_header,
+                "source": f"bio2zarr-{provenance.__version__}",
+            }
+        )
+        # Doing this syncronously - this is fine surely
+        self.encode_samples(root)
+        self.encode_filter_id(root)
+        self.encode_contig_id(root)
+
+        self.wip_path.mkdir()
+        self.arrays_path.mkdir()
+        self.partitions_path.mkdir()
+        root = zarr.open(
+            store=self.arrays_path, mode="a", **zarr_utils.ZARR_FORMAT_KWARGS
+        )
+
+        total_chunks = 0
+        for field in self.schema.fields:
+            a = self.init_array(root, field, partitions[-1].stop)
+            total_chunks += a.nchunks
+
+        logger.info("Writing WIP metadata")
+        with open(self.wip_path / "metadata.json", "w") as f:
+            json.dump(self.metadata.asdict(), f, indent=4)
+
+        return VcfZarrWriteSummary(
+            num_variants=self.icf.num_records,
+            num_samples=self.icf.num_samples,
+            num_partitions=self.num_partitions,
+            num_chunks=total_chunks,
+            max_encoding_memory=core.display_size(self.get_max_encoding_memory()),
+        )
+
+    def encode_samples(self, root):
+        if self.schema.samples != self.icf.metadata.samples:
+            raise ValueError("Subsetting or reordering samples not supported currently")
+        array = root.array(
+            "sample_id",
+            data=[sample.id for sample in self.schema.samples],
+            shape=len(self.schema.samples),
+            dtype="str",
+            compressor=schema.DEFAULT_ZARR_COMPRESSOR,
+            chunks=(self.schema.samples_chunk_size,),
+        )
+        array.attrs["_ARRAY_DIMENSIONS"] = ["samples"]
+        logger.debug("Samples done")
+
+    def encode_contig_id(self, root):
+        array = root.array(
+            "contig_id",
+            data=[contig.id for contig in self.schema.contigs],
+            shape=len(self.schema.contigs),
+            dtype="str",
+            compressor=schema.DEFAULT_ZARR_COMPRESSOR,
+        )
+        array.attrs["_ARRAY_DIMENSIONS"] = ["contigs"]
+        if all(contig.length is not None for contig in self.schema.contigs):
+            array = root.array(
+                "contig_length",
+                data=[contig.length for contig in self.schema.contigs],
+                shape=len(self.schema.contigs),
+                dtype=np.int64,
+                compressor=schema.DEFAULT_ZARR_COMPRESSOR,
+            )
+            array.attrs["_ARRAY_DIMENSIONS"] = ["contigs"]
+
+    def encode_filter_id(self, root):
+        # TODO need a way to store description also
+        # https://github.com/sgkit-dev/vcf-zarr-spec/issues/19
+        array = root.array(
+            "filter_id",
+            data=[filt.id for filt in self.schema.filters],
+            shape=len(self.schema.filters),
+            dtype="str",
+            compressor=schema.DEFAULT_ZARR_COMPRESSOR,
+        )
+        array.attrs["_ARRAY_DIMENSIONS"] = ["filters"]
+
+    def init_array(self, root, array_spec, variants_dim_size):
+        kwargs = dict(zarr_utils.ZARR_FORMAT_KWARGS)
+        filters = [numcodecs.get_codec(filt) for filt in array_spec.filters]
+        if array_spec.dtype == "O":
+            if zarr_utils.zarr_v3():
+                filters = [*list(filters), numcodecs.VLenUTF8()]
+            else:
+                kwargs["object_codec"] = numcodecs.VLenUTF8()
+
+        if not zarr_utils.zarr_v3():
+            kwargs["dimension_separator"] = self.metadata.dimension_separator
+
+        shape = list(array_spec.shape)
+        # Truncate the variants dimension is max_variant_chunks was specified
+        shape[0] = variants_dim_size
+        a = root.empty(
+            name=array_spec.name,
+            shape=shape,
+            chunks=array_spec.chunks,
+            dtype=array_spec.dtype,
+            compressor=numcodecs.get_codec(array_spec.compressor),
+            filters=filters,
+            **kwargs,
+        )
+        a.attrs.update(
+            {
+                "description": array_spec.description,
+                # Dimension names are part of the spec in Zarr v3
+                "_ARRAY_DIMENSIONS": array_spec.dimensions,
+            }
+        )
+        logger.debug(f"Initialised {a}")
+        return a
+
+    #######################
+    # encode_partition
+    #######################
+
+    def load_metadata(self):
+        if self.metadata is None:
+            with open(self.wip_path / "metadata.json") as f:
+                self.metadata = VcfZarrWriterMetadata.fromdict(json.load(f))
+            self.icf = icf.IntermediateColumnarFormat(self.metadata.icf_path)
+
+    def partition_path(self, partition_index):
+        return self.partitions_path / f"p{partition_index}"
+
+    def wip_partition_path(self, partition_index):
+        return self.partitions_path / f"wip_p{partition_index}"
+
+    def wip_partition_array_path(self, partition_index, name):
+        return self.wip_partition_path(partition_index) / name
+
+    def partition_array_path(self, partition_index, name):
+        return self.partition_path(partition_index) / name
+
+    def encode_partition(self, partition_index):
+        self.load_metadata()
+        if partition_index < 0 or partition_index >= self.num_partitions:
+            raise ValueError("Partition index not in the valid range")
+        partition_path = self.wip_partition_path(partition_index)
+        partition_path.mkdir(exist_ok=True)
+        logger.info(f"Encoding partition {partition_index} to {partition_path}")
+
+        self.encode_id_partition(partition_index)
+        self.encode_filters_partition(partition_index)
+        self.encode_contig_partition(partition_index)
+        self.encode_alleles_partition(partition_index)
+        for array_spec in self.schema.fields:
+            if array_spec.vcf_field is not None:
+                self.encode_array_partition(array_spec, partition_index)
+        if self.has_genotypes():
+            self.encode_genotypes_partition(partition_index)
+            self.encode_genotype_mask_partition(partition_index)
+        if self.has_local_alleles():
+            self.encode_local_alleles_partition(partition_index)
+            self.encode_local_allele_fields_partition(partition_index)
+
+        final_path = self.partition_path(partition_index)
+        logger.info(f"Finalising {partition_index} at {final_path}")
+        if final_path.exists():
+            logger.warning(f"Removing existing partition at {final_path}")
+            shutil.rmtree(final_path)
+        os.rename(partition_path, final_path)
+
+    def init_partition_array(self, partition_index, name):
+        field_map = self.schema.field_map()
+        array_spec = field_map[name]
+        # Create an empty array like the definition
+        src = self.arrays_path / array_spec.name
+        # Overwrite any existing WIP files
+        wip_path = self.wip_partition_array_path(partition_index, array_spec.name)
+        shutil.copytree(src, wip_path, dirs_exist_ok=True)
+        array = zarr.open_array(store=wip_path, mode="a")
+        partition = self.metadata.partitions[partition_index]
+        ba = core.BufferedArray(array, partition.start, name)
+        logger.info(
+            f"Start partition {partition_index} array {name} <{array.dtype}> "
+            f"{array.shape} @ {wip_path}"
+        )
+        return ba
+
+    def finalise_partition_array(self, partition_index, buffered_array):
+        buffered_array.flush()
+        logger.info(
+            f"Completed partition {partition_index} array {buffered_array.name} "
+            f"max_memory={core.display_size(buffered_array.max_buff_size)}"
+        )
+
+    def encode_array_partition(self, array_spec, partition_index):
+        partition = self.metadata.partitions[partition_index]
+        ba = self.init_partition_array(partition_index, array_spec.name)
+        source_field = self.icf.fields[array_spec.vcf_field]
+        sanitiser = source_field.sanitiser_factory(ba.buff.shape)
+
+        for value in source_field.iter_values(partition.start, partition.stop):
+            # We write directly into the buffer in the sanitiser function
+            # to make it easier to reason about dimension padding
+            j = ba.next_buffer_row()
+            sanitiser(ba.buff, j, value)
+        self.finalise_partition_array(partition_index, ba)
+
+    def encode_genotypes_partition(self, partition_index):
+        partition = self.metadata.partitions[partition_index]
+        gt = self.init_partition_array(partition_index, "call_genotype")
+        gt_phased = self.init_partition_array(partition_index, "call_genotype_phased")
+
+        source_field = self.icf.fields["FORMAT/GT"]
+        for value in source_field.iter_values(partition.start, partition.stop):
+            j = gt.next_buffer_row()
+            icf.sanitise_value_int_2d(
+                gt.buff, j, value[:, :-1] if value is not None else None
+            )
+            j = gt_phased.next_buffer_row()
+            icf.sanitise_value_int_1d(
+                gt_phased.buff, j, value[:, -1] if value is not None else None
+            )
+
+        self.finalise_partition_array(partition_index, gt)
+        self.finalise_partition_array(partition_index, gt_phased)
+
+    def encode_genotype_mask_partition(self, partition_index):
+        partition = self.metadata.partitions[partition_index]
+        gt_mask = self.init_partition_array(partition_index, "call_genotype_mask")
+        # Read back in the genotypes so we can compute the mask
+        gt_array = zarr.open_array(
+            store=self.wip_partition_array_path(partition_index, "call_genotype"),
+            mode="r",
+        )
+        for genotypes in core.first_dim_slice_iter(
+            gt_array, partition.start, partition.stop
+        ):
+            # TODO check is this the correct semantics when we are padding
+            # with mixed ploidies?
+            j = gt_mask.next_buffer_row()
+            gt_mask.buff[j] = genotypes < 0
+        self.finalise_partition_array(partition_index, gt_mask)
+
+    def encode_local_alleles_partition(self, partition_index):
+        partition = self.metadata.partitions[partition_index]
+        call_LA = self.init_partition_array(partition_index, "call_LA")
+
+        gt_array = zarr.open_array(
+            store=self.wip_partition_array_path(partition_index, "call_genotype"),
+            mode="r",
+        )
+        for genotypes in core.first_dim_slice_iter(
+            gt_array, partition.start, partition.stop
+        ):
+            la = vcz.compute_la_field(genotypes)
+            j = call_LA.next_buffer_row()
+            call_LA.buff[j] = la
+        self.finalise_partition_array(partition_index, call_LA)
+
+    def encode_local_allele_fields_partition(self, partition_index):
+        partition = self.metadata.partitions[partition_index]
+        la_array = zarr.open_array(
+            store=self.wip_partition_array_path(partition_index, "call_LA"),
+            mode="r",
+        )
+        # We got through the localisable fields one-by-one so that we don't need to
+        # keep several large arrays in memory at once for each partition.
+        field_map = self.schema.field_map()
+        for descriptor in localisable_fields:
+            if descriptor.array_name not in field_map:
+                continue
+            assert field_map[descriptor.array_name].vcf_field is None
+
+            buff = self.init_partition_array(partition_index, descriptor.array_name)
+            source = self.icf.fields[descriptor.vcf_field].iter_values(
+                partition.start, partition.stop
+            )
+            for la in core.first_dim_slice_iter(
+                la_array, partition.start, partition.stop
+            ):
+                raw_value = next(source)
+                value = descriptor.sanitise(raw_value, 2, raw_value.dtype)
+                j = buff.next_buffer_row()
+                buff.buff[j] = descriptor.convert(value, la)
+            self.finalise_partition_array(partition_index, buff)
+
+    def encode_alleles_partition(self, partition_index):
+        alleles = self.init_partition_array(partition_index, "variant_allele")
+        partition = self.metadata.partitions[partition_index]
+        ref_field = self.icf.fields["REF"]
+        alt_field = self.icf.fields["ALT"]
+
+        for ref, alt in zip(
+            ref_field.iter_values(partition.start, partition.stop),
+            alt_field.iter_values(partition.start, partition.stop),
+        ):
+            j = alleles.next_buffer_row()
+            alleles.buff[j, :] = constants.STR_FILL
+            alleles.buff[j, 0] = ref[0]
+            alleles.buff[j, 1 : 1 + len(alt)] = alt
+        self.finalise_partition_array(partition_index, alleles)
+
+    def encode_id_partition(self, partition_index):
+        vid = self.init_partition_array(partition_index, "variant_id")
+        vid_mask = self.init_partition_array(partition_index, "variant_id_mask")
+        partition = self.metadata.partitions[partition_index]
+        field = self.icf.fields["ID"]
+
+        for value in field.iter_values(partition.start, partition.stop):
+            j = vid.next_buffer_row()
+            k = vid_mask.next_buffer_row()
+            assert j == k
+            if value is not None:
+                vid.buff[j] = value[0]
+                vid_mask.buff[j] = False
+            else:
+                vid.buff[j] = constants.STR_MISSING
+                vid_mask.buff[j] = True
+
+        self.finalise_partition_array(partition_index, vid)
+        self.finalise_partition_array(partition_index, vid_mask)
+
+    def encode_filters_partition(self, partition_index):
+        lookup = {filt.id: index for index, filt in enumerate(self.schema.filters)}
+        var_filter = self.init_partition_array(partition_index, "variant_filter")
+        partition = self.metadata.partitions[partition_index]
+
+        field = self.icf.fields["FILTERS"]
+        for value in field.iter_values(partition.start, partition.stop):
+            j = var_filter.next_buffer_row()
+            var_filter.buff[j] = False
+            for f in value:
+                try:
+                    var_filter.buff[j, lookup[f]] = True
+                except KeyError:
+                    raise ValueError(
+                        f"Filter '{f}' was not defined in the header."
+                    ) from None
+
+        self.finalise_partition_array(partition_index, var_filter)
+
+    def encode_contig_partition(self, partition_index):
+        lookup = {contig.id: index for index, contig in enumerate(self.schema.contigs)}
+        contig = self.init_partition_array(partition_index, "variant_contig")
+        partition = self.metadata.partitions[partition_index]
+        field = self.icf.fields["CHROM"]
+
+        for value in field.iter_values(partition.start, partition.stop):
+            j = contig.next_buffer_row()
+            # Note: because we are using the indexes to define the lookups
+            # and we always have an index, it seems that we the contig lookup
+            # will always succeed. However, if anyone ever does hit a KeyError
+            # here, please do open an issue with a reproducible example!
+            contig.buff[j] = lookup[value[0]]
+
+        self.finalise_partition_array(partition_index, contig)
+
+    #######################
+    # finalise
+    #######################
+
+    def finalise_array(self, name):
+        logger.info(f"Finalising {name}")
+        final_path = self.path / name
+        if final_path.exists():
+            # NEEDS TEST
+            raise ValueError(f"Array {name} already exists")
+        for partition in range(self.num_partitions):
+            # Move all the files in partition dir to dest dir
+            src = self.partition_array_path(partition, name)
+            if not src.exists():
+                # Needs test
+                raise ValueError(f"Partition {partition} of {name} does not exist")
+            dest = self.arrays_path / name
+            # This is Zarr v2 specific. Chunks in v3 with start with "c" prefix.
+            chunk_files = [
+                path for path in src.iterdir() if not path.name.startswith(".")
+            ]
+            # TODO check for a count of then number of files. If we require a
+            # dimension_separator of "/" then we could make stronger assertions
+            # here, as we'd always have num_variant_chunks
+            logger.debug(
+                f"Moving {len(chunk_files)} chunks for {name} partition {partition}"
+            )
+            for chunk_file in chunk_files:
+                os.rename(chunk_file, dest / chunk_file.name)
+        # Finally, once all the chunks have moved into the arrays dir,
+        # we move it out of wip
+        os.rename(self.arrays_path / name, self.path / name)
+        core.update_progress(1)
+
+    def finalise(self, show_progress=False):
+        self.load_metadata()
+
+        logger.info(f"Scanning {self.num_partitions} partitions")
+        missing = []
+        # TODO may need a progress bar here
+        for partition_id in range(self.num_partitions):
+            if not self.partition_path(partition_id).exists():
+                missing.append(partition_id)
+        if len(missing) > 0:
+            raise FileNotFoundError(f"Partitions not encoded: {missing}")
+
+        progress_config = core.ProgressConfig(
+            total=len(self.schema.fields),
+            title="Finalise",
+            units="array",
+            show=show_progress,
+        )
+        # NOTE: it's not clear that adding more workers will make this quicker,
+        # as it's just going to be causing contention on the file system.
+        # Something to check empirically in some deployments.
+        # FIXME we're just using worker_processes=0 here to hook into the
+        # SynchronousExecutor which is intended for testing purposes so
+        # that we get test coverage. Should fix this either by allowing
+        # for multiple workers, or making a standard wrapper for tqdm
+        # that allows us to have a consistent look and feel.
+        with core.ParallelWorkManager(0, progress_config) as pwm:
+            for field in self.schema.fields:
+                pwm.submit(self.finalise_array, field.name)
+        logger.debug(f"Removing {self.wip_path}")
+        shutil.rmtree(self.wip_path)
+        logger.info("Consolidating Zarr metadata")
+        zarr.consolidate_metadata(self.path)
+
+    #######################
+    # index
+    #######################
+
+    def create_index(self):
+        """Create an index to support efficient region queries."""
+
+        root = zarr.open_group(store=self.path, mode="r+")
+
+        contig = root["variant_contig"]
+        pos = root["variant_position"]
+        length = root["variant_length"]
+
+        assert contig.cdata_shape == pos.cdata_shape
+
+        index = []
+
+        logger.info("Creating region index")
+        for v_chunk in range(pos.cdata_shape[0]):
+            c = contig.blocks[v_chunk]
+            p = pos.blocks[v_chunk]
+            e = p + length.blocks[v_chunk] - 1
+
+            # create a row for each contig in the chunk
+            d = np.diff(c, append=-1)
+            c_start_idx = 0
+            for c_end_idx in np.nonzero(d)[0]:
+                assert c[c_start_idx] == c[c_end_idx]
+                index.append(
+                    (
+                        v_chunk,  # chunk index
+                        c[c_start_idx],  # contig ID
+                        p[c_start_idx],  # start
+                        p[c_end_idx],  # end
+                        np.max(e[c_start_idx : c_end_idx + 1]),  # max end
+                        c_end_idx - c_start_idx + 1,  # num records
+                    )
+                )
+                c_start_idx = c_end_idx + 1
+
+        index = np.array(index, dtype=pos.dtype)
+        kwargs = {}
+        if not zarr_utils.zarr_v3():
+            kwargs["dimension_separator"] = self.metadata.dimension_separator
+        array = root.array(
+            "region_index",
+            data=index,
+            shape=index.shape,
+            chunks=index.shape,
+            dtype=index.dtype,
+            compressor=numcodecs.Blosc("zstd", clevel=9, shuffle=0),
+            fill_value=None,
+            **kwargs,
+        )
+        array.attrs["_ARRAY_DIMENSIONS"] = [
+            "region_index_values",
+            "region_index_fields",
+        ]
+
+        logger.info("Consolidating Zarr metadata")
+        zarr.consolidate_metadata(self.path)
+
+    ######################
+    # encode_all_partitions
+    ######################
+
+    def get_max_encoding_memory(self):
+        """
+        Return the approximate maximum memory used to encode a variant chunk.
+        """
+        max_encoding_mem = 0
+        for array_spec in self.schema.fields:
+            max_encoding_mem = max(max_encoding_mem, array_spec.variant_chunk_nbytes)
+        gt_mem = 0
+        if self.has_genotypes:
+            gt_mem = sum(
+                field.variant_chunk_nbytes
+                for field in self.schema.fields
+                if field.name.startswith("call_genotype")
+            )
+        return max(max_encoding_mem, gt_mem)
+
+    def encode_all_partitions(
+        self, *, worker_processes=1, show_progress=False, max_memory=None
+    ):
+        max_memory = core.parse_max_memory(max_memory)
+        self.load_metadata()
+        num_partitions = self.num_partitions
+        per_worker_memory = self.get_max_encoding_memory()
+        logger.info(
+            f"Encoding Zarr over {num_partitions} partitions with "
+            f"{worker_processes} workers and {core.display_size(per_worker_memory)} "
+            "per worker"
+        )
+        # Each partition requires per_worker_memory bytes, so to prevent more that
+        # max_memory being used, we clamp the number of workers
+        max_num_workers = max_memory // per_worker_memory
+        if max_num_workers < worker_processes:
+            logger.warning(
+                f"Limiting number of workers to {max_num_workers} to "
+                "keep within specified memory budget of "
+                f"{core.display_size(max_memory)}"
+            )
+        if max_num_workers <= 0:
+            raise ValueError(
+                f"Insufficient memory to encode a partition:"
+                f"{core.display_size(per_worker_memory)} > "
+                f"{core.display_size(max_memory)}"
+            )
+        num_workers = min(max_num_workers, worker_processes)
+
+        total_bytes = 0
+        for array_spec in self.schema.fields:
+            # Open the array definition to get the total size
+            total_bytes += zarr.open(self.arrays_path / array_spec.name).nbytes
+
+        progress_config = core.ProgressConfig(
+            total=total_bytes,
+            title="Encode",
+            units="B",
+            show=show_progress,
+        )
+        with core.ParallelWorkManager(num_workers, progress_config) as pwm:
+            for partition_index in range(num_partitions):
+                pwm.submit(self.encode_partition, partition_index)
+
+
+VZW_METADATA_FORMAT_VERSION = "0.1"
+
+
+@dataclasses.dataclass
+class VcfZarrWriterMetadata(core.JsonDataclass):
+    format_version: str
+    icf_path: str
+    schema: schema.VcfZarrSchema
+    dimension_separator: str
+    partitions: list
+    provenance: dict
+
+    @staticmethod
+    def fromdict(d):
+        if d["format_version"] != VZW_METADATA_FORMAT_VERSION:
+            raise ValueError(
+                "VcfZarrWriter format version mismatch: "
+                f"{d['format_version']} != {VZW_METADATA_FORMAT_VERSION}"
+            )
+        ret = VcfZarrWriterMetadata(**d)
+        ret.schema = schema.VcfZarrSchema.fromdict(ret.schema)
+        ret.partitions = [VcfZarrPartition(**p) for p in ret.partitions]
+        return ret
+
+
+@dataclasses.dataclass
+class VcfZarrPartition:
+    start: int
+    stop: int
+
+    @staticmethod
+    def generate_partitions(num_records, chunk_size, num_partitions, max_chunks=None):
+        num_chunks = int(np.ceil(num_records / chunk_size))
+        if max_chunks is not None:
+            num_chunks = min(num_chunks, max_chunks)
+        partitions = []
+        splits = np.array_split(np.arange(num_chunks), min(num_partitions, num_chunks))
+        for chunk_slice in splits:
+            start_chunk = int(chunk_slice[0])
+            stop_chunk = int(chunk_slice[-1]) + 1
+            start_index = start_chunk * chunk_size
+            stop_index = min(stop_chunk * chunk_size, num_records)
+            partitions.append(VcfZarrPartition(start_index, stop_index))
+        return partitions
+
+
+class VcfZarr:
+    def __init__(self, path):
+        if not (path / ".zmetadata").exists():
+            raise ValueError("Not in VcfZarr format")  # NEEDS TEST
+        self.path = path
+        self.root = zarr.open(path, mode="r")
+
+    def summary_table(self):
+        data = []
+        arrays = [(core.du(self.path / a.basename), a) for _, a in self.root.arrays()]
+        arrays.sort(key=lambda x: x[0])
+        for stored, array in reversed(arrays):
+            d = {
+                "name": array.name,
+                "dtype": str(array.dtype),
+                "stored": core.display_size(stored),
+                "size": core.display_size(array.nbytes),
+                "ratio": core.display_number(array.nbytes / stored),
+                "nchunks": str(array.nchunks),
+                "chunk_size": core.display_size(array.nbytes / array.nchunks),
+                "avg_chunk_stored": core.display_size(int(stored / array.nchunks)),
+                "shape": str(array.shape),
+                "chunk_shape": str(array.chunks),
+                "compressor": str(array.compressor),
+                "filters": str(array.filters),
+            }
+            data.append(d)
+        return data
+
+
+import dataclasses
+import logging
+import multiprocessing
+
+from tqdm.auto import tqdm
+
+from bio2zarr import schema
+from bio2zarr.zarr_utils import ZARR_FORMAT_KWARGS
+
+logger = logging.getLogger(__name__)
+
+# Keep existing VcfZarrWriter class...
+
+
+class GenericZarrWriter:
+    """
+    A general zarr writer that can handle data from multiple sources (VCF, PLINK, etc.)
+    using appropriate data adapters.
+    """
+
+    def __init__(self, path):
+        self.path = pathlib.Path(path)
+        self.zarr_root = None
+        self.schema = None
+        self.field_map = None
+        self.partitioner = None
+
+    def init_from_schema(self, schema_instance, dimension_separator="/"):
+        """
+        Initialize the zarr store using a schema that was generated for a specific data source.
+        """
+        self.schema = schema_instance
+        self.schema.validate()
+        self.field_map = self.schema.field_map()
+
+        # Create the zarr store
+        self.zarr_root = zarr.open(store=self.path, mode="w", **ZARR_FORMAT_KWARGS)
+        if dimension_separator is not None:
+            self.zarr_root.attrs["_ARRAY_DIMENSIONS"] = dimension_separator
+
+        # Create arrays based on the schema
+        for field in self.schema.fields:
+            self._create_array(field, dimension_separator)
+
+        # Store schema as metadata
+        self.zarr_root.attrs["bio2zarr_schema"] = json.loads(self.schema.asjson())
+        return self
+
+    def _create_array(self, field_spec, dimension_separator):
+        """
+        Create a zarr array based on the provided field specification.
+        """
+        array = self.zarr_root.create_dataset(
+            name=field_spec.name,
+            shape=field_spec.shape,
+            chunks=field_spec.chunks,
+            dtype=field_spec.dtype,
+            compressor=numcodecs.get_codec(field_spec.compressor),
+            filters=field_spec.filters or None,
+        )
+        array.attrs["_ARRAY_DIMENSIONS"] = (
+            dimension_separator.join(field_spec.dimensions)
+            if dimension_separator is not None
+            else field_spec.dimensions
+        )
+        array.attrs["description"] = field_spec.description
+        if field_spec.vcf_field is not None:
+            array.attrs["vcf_field"] = field_spec.vcf_field
+
+    def encode_data(self, data_adapter, worker_processes=1, show_progress=False):
+        """
+        Encode data from a source adapter into the zarr store.
+
+        The data_adapter should implement methods to provide data for each array
+        defined in the schema.
+        """
+        # Encode basic metadata
+        self._encode_metadata(data_adapter)
+
+        # Encode genotypes in chunks/partitions
+        variants_chunk_size = self.schema.variants_chunk_size
+        n_variants = self.field_map["call_genotype"].shape[0]
+
+        # Create a list of variant slice ranges to process
+        slices = []
+        for start in range(0, n_variants, variants_chunk_size):
+            stop = min(start + variants_chunk_size, n_variants)
+            slices.append((start, stop))
+
+        # Process slices
+        if show_progress:
+            slices_iter = tqdm(slices, desc="Encoding genotypes")
+        else:
+            slices_iter = slices
+
+        if worker_processes > 1:
+            with multiprocessing.Pool(worker_processes) as pool:
+                pool.starmap(
+                    self._encode_genotypes_slice,
+                    [(data_adapter, start, stop) for start, stop in slices_iter],
+                )
+        else:
+            for start, stop in slices_iter:
+                self._encode_genotypes_slice(data_adapter, start, stop)
+
+    def _encode_metadata(self, data_adapter):
+        """
+        Encode metadata like sample IDs, variant positions, etc.
+        """
+        # Encode any metadata provided by the adapter
+        # This will be source-specific and depends on what methods the adapter implements
+        if hasattr(data_adapter, "get_sample_ids"):
+            sample_ids = data_adapter.get_sample_ids()
+            if "sample_id" in self.zarr_root:
+                self.zarr_root["sample_id"][:] = sample_ids
+
+        if hasattr(data_adapter, "get_variant_positions"):
+            positions = data_adapter.get_variant_positions()
+            if "variant_position" in self.zarr_root:
+                self.zarr_root["variant_position"][:] = positions
+
+        if hasattr(data_adapter, "get_variant_alleles"):
+            alleles = data_adapter.get_variant_alleles()
+            if "variant_allele" in self.zarr_root:
+                self.zarr_root["variant_allele"][:] = alleles
+
+    def _encode_genotypes_slice(self, data_adapter, start, stop):
+        """
+        Encode a slice of genotypes from the data adapter.
+        """
+        # Get the genotypes from the adapter
+        if hasattr(data_adapter, "get_genotypes_slice"):
+            genotype_data = data_adapter.get_genotypes_slice(start, stop)
+
+            # Write each array
+            for array_name, data in genotype_data.items():
+                if array_name in self.zarr_root:
+                    if len(data.shape) == 3:  # Genotypes with ploidy
+                        self.zarr_root[array_name][start:stop, :, :] = data
+                    elif len(data.shape) == 2:  # Phased info
+                        self.zarr_root[array_name][start:stop, :] = data
+
+    def finalise(self, show_progress=False):
+        """
+        Finalize the zarr store by consolidating metadata.
+        """
+        zarr.consolidate_metadata(self.zarr_root.store)
+        logger.info(f"Zarr dataset created at: {self.path}")
